Code refactoring based on SonarLint

+++ src/external_measures/statistical_hypothesis/FowlkesMallowsIndex.java
@@ -7,20 +7,20 @@ import interfaces.Hypotheses;
 public class FowlkesMallowsIndex extends CommonQualityMeasure {
 	private Hypotheses hypothesesCalculator;
 
-	private FowlkesMallowsIndex() {}
-	
-	public FowlkesMallowsIndex(Hypotheses hypothesesCalculator)
-	{
+	private FowlkesMallowsIndex() {
+	}
+
+	public FowlkesMallowsIndex(Hypotheses hypothesesCalculator) {
 		this.hypothesesCalculator = hypothesesCalculator;
 	}
-	
+
 	@Override
 	public double getMeasure(Hierarchy h) {
 		this.hypothesesCalculator.calculate(h);
-		double numerator = this.hypothesesCalculator.getTP()*this.hypothesesCalculator.getTP();
+		double numerator = this.hypothesesCalculator.getTP() * (double) this.hypothesesCalculator.getTP();
 		double denominator = (this.hypothesesCalculator.getTP() + this.hypothesesCalculator.getFP())
-                * (this.hypothesesCalculator.getTP() + this.hypothesesCalculator.getFN());
-        return Math.sqrt(numerator/denominator);
+				* (double) (this.hypothesesCalculator.getTP() + this.hypothesesCalculator.getFN());
+		return Math.sqrt(numerator / denominator);
 	}
 
 	@Override
